Image management and custom exceptions implemented

+++ Backend/src/main/java/com/pg/mbti/services/FriendshipsService.java
@@ -2,6 +2,9 @@ package com.pg.mbti.services;
 
 import com.pg.mbti.entity.Friendship;
 import com.pg.mbti.entity.User;
+import com.pg.mbti.exceptions.FriendshipAlreadyExistsException;
+import com.pg.mbti.exceptions.FriendshipNotFoundException;
+import com.pg.mbti.exceptions.ResourceNotFoundException;
 import com.pg.mbti.repositories.FriendshipsRepository;
 import com.pg.mbti.repositories.UsersRepository;
 import lombok.AllArgsConstructor;
@@ -13,7 +16,7 @@ import java.util.List;
 @AllArgsConstructor
 public class FriendshipsService {
     FriendshipsRepository friendshipRepository;
-    UsersRepository  usersRepository;
+    UsersRepository usersRepository;
 
     public List<Friendship> getAllFriendships() {
         return friendshipRepository.findAll();
@@ -27,26 +30,41 @@ public class FriendshipsService {
         return friendshipRepository.getFriendshipsByNickname(name);
     }
 
-    private User findUserByNickname(String Nickname) {
+    private User findUserByNickname(String nickname) {
         return usersRepository
-                .findByNickname(Nickname)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .findByNickname(nickname)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found: " + nickname));
     }
 
     public void createFriendship(String currentNickname, String newFriendNickname) {
         User sender = findUserByNickname(currentNickname);
         User receiver = findUserByNickname(newFriendNickname);
+
+        if (friendshipRepository.existsByFriends(currentNickname, newFriendNickname)) {
+            throw new FriendshipAlreadyExistsException("Friendship already exists");
+        }
+
         friendshipRepository.save(new Friendship(sender, receiver));
     }
 
     public void acceptFriendship(String currentNickname, String newFriendNickname) {
         Friendship friendship = friendshipRepository.findByFriends(currentNickname, newFriendNickname);
+
+        if (friendship == null) {
+            throw new FriendshipNotFoundException("Friendship request not found");
+        }
+
         friendship.setPending(false);
         friendshipRepository.save(friendship);
     }
 
     public void deleteFriendship(String currentNickname, String newFriendNickname) {
         Friendship friendship = friendshipRepository.findByFriends(currentNickname, newFriendNickname);
+
+        if (friendship == null) {
+            throw new FriendshipNotFoundException("Friendship not found");
+        }
+
         friendshipRepository.delete(friendship);
     }
 }
